pyproject.toml
[project]
name = "latest-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.3",
    "fpdf2>=2.8.7",
    "google-genai>=1.70.0",
    "opencv-python-headless>=4.13.0.92",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.1",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.2.2",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "werkzeug>=3.1.7",
]
